# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/places/autocomplete:
    get:
      operationId: autocompletePlaces
      tags: [places]
      summary: Search Google Places by free-text business name/address
      description: >
        Public (unauthenticated) — used on the pre-registration business
        search step. Proxies Google Places Autocomplete so the API key never
        reaches the browser.
      parameters:
        - name: input
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Autocomplete suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaceAutocompleteResult"
        "400":
          description: Missing input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Upstream Google Places error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/places/details/{placeId}:
    get:
      operationId: getPlaceDetails
      tags: [places]
      summary: Fetch full details for a selected Google Place
      description: >
        Public (unauthenticated) — used to pre-fill the registration form
        once the owner picks a suggestion.
      parameters:
        - name: placeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Place details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaceDetails"
        "502":
          description: Upstream Google Places error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/places/photo:
    get:
      operationId: getPlacePhoto
      tags: [places]
      summary: Proxy a Google Place photo so the API key stays server-side
      description: >
        Public (unauthenticated). `name` must be the raw photo resource name
        returned as `photoName` on PlaceDetails (e.g. "places/X/photos/Y").
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Image bytes streamed with correct Content-Type
          content:
            image/*:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid photo reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Upstream Google Places error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/public/review/{businessSlug}/{campaignSlug}:
    get:
      operationId: getPublicReviewPage
      tags: [public-review]
      summary: Fetch the data needed to render a public review page
      description: >
        Public (unauthenticated). Returns 404 if the business or campaign
        doesn't exist, isn't ACTIVE, or is archived/deleted, so unpublished
        campaigns are never reachable by guessing a URL.
      parameters:
        - name: businessSlug
          in: path
          required: true
          schema:
            type: string
        - name: campaignSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public review page data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicReviewPageResult"
        "404":
          description: Not found or not active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/public/review/{businessSlug}/{campaignSlug}/generate:
    post:
      operationId: generatePublicReview
      tags: [public-review]
      summary: Generate (or regenerate) an AI review draft for a customer session
      description: >
        Public (unauthenticated). Capped at 3 generations per sessionId to
        bound AI spend per customer visit; the 4th call returns 429.
      parameters:
        - name: businessSlug
          in: path
          required: true
          schema:
            type: string
        - name: campaignSlug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicGenerateReviewInput"
      responses:
        "200":
          description: Generated review text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicGenerateReviewResult"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found or not active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Regeneration limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: AI provider error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: string
        message:
          type: string
      required:
        - success
        - code
        - message
...
    KeywordCategory:
      type: string
      enum: [PRODUCT_SERVICE, EXPERIENCE]
...
    PlaceAutocompleteSuggestion:
      type: object
      properties:
        placeId:
          type: string
        mainText:
          type: string
        secondaryText:
          type: ["string", "null"]
        description:
          type: string
      required:
        - placeId
        - mainText
        - secondaryText
        - description
    PlaceAutocompleteResult:
      type: object
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/PlaceAutocompleteSuggestion"
      required:
        - suggestions
    PlaceDetails:
      type: object
      properties:
        placeId:
          type: string
        name:
          type: string
        category:
          type: ["string", "null"]
        formattedAddress:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        rating:
          type: ["number", "null"]
        userRatingCount:
          type: ["integer", "null"]
        photoName:
          type: ["string", "null"]
      required:
        - placeId
        - name
        - category
        - formattedAddress
        - phone
        - website
        - latitude
        - longitude
        - rating
        - userRatingCount
        - photoName
    PublicBusinessSummary:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        logoUrl:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        brandColor:
          type: ["string", "null"]
        welcomeMessage:
          type: ["string", "null"]
      required:
        - name
        - category
        - logoUrl
        - coverImageUrl
        - brandColor
        - welcomeMessage
    PublicCampaignSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      required:
        - id
        - name
    PublicKeywordSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        category:
          $ref: "#/components/schemas/KeywordCategory"
      required:
        - id
        - label
        - category
    PublicReviewPageResult:
      type: object
      properties:
        business:
          $ref: "#/components/schemas/PublicBusinessSummary"
        campaign:
          $ref: "#/components/schemas/PublicCampaignSummary"
        keywords:
          type: array
          items:
            $ref: "#/components/schemas/PublicKeywordSummary"
        googleReviewUrl:
          type: ["string", "null"]
      required:
        - business
        - campaign
        - keywords
        - googleReviewUrl
    PublicGenerateReviewInput:
      type: object
      properties:
        sessionId:
          type: string
          minLength: 1
        keywords:
          type: array
          items:
            type: string
          minItems: 1
      required:
        - sessionId
        - keywords
    PublicGenerateReviewResult:
      type: object
      properties:
        reviewText:
          type: string
        remainingGenerations:
          type: integer
        maxGenerations:
          type: integer
      required:
        - reviewText
        - remainingGenerations
        - maxGenerations
...