# lib/api-zod/src/generated/types/placeAutocompleteSuggestion.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlaceAutocompleteSuggestion {
  placeId: string;
  mainText: string;
  /** @nullable */
  secondaryText: string | null;
  description: string;
}