# lib/api-zod/src/generated/types/publicBusinessSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PublicBusinessSummary {
  name: string;
  category: string;
  /** @nullable */
  logoUrl: string | null;
  /** @nullable */
  coverImageUrl: string | null;
  /** @nullable */
  brandColor: string | null;
  /** @nullable */
  welcomeMessage: string | null;
}